microservicio/infraestructura/src/test/java/com/ceiba/factura/controlador/ConsultaControladorFacturaTest.java
package com.ceiba.factura.controlador;

import com.ceiba.ApplicationMock;
import org.junit.jupiter.api.DisplayName;
import org.junit.jupiter.api.Test;
import org.junit.jupiter.api.extension.ExtendWith;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
import org.springframework.http.MediaType;
import org.springframework.test.annotation.DirtiesContext;
import org.springframework.test.context.ContextConfiguration;
import org.springframework.test.context.junit.jupiter.SpringExtension;
import org.springframework.test.web.servlet.MockMvc;

import java.time.LocalDateTime;

import static org.hamcrest.Matchers.hasSize;
import static org.hamcrest.core.Is.is;
import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.get;
import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.jsonPath;
import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;

@ExtendWith(SpringExtension.class)
@WebMvcTest(ConsultaControladorFactura.class)
@ContextConfiguration(classes= ApplicationMock.class)
@DirtiesContext(classMode = DirtiesContext.ClassMode.AFTER_EACH_TEST_METHOD)
public class ConsultaControladorFacturaTest {
    @Autowired
    private MockMvc mocMvc;

    @Test
    @DisplayName("Deberia listar las facturas en memoria")
    void deberiaListarFactura() throws Exception {
        // arrange
        String fechaPartida [] = String.valueOf(LocalDateTime.now()).split("T");
        // act - assert
        mocMvc.perform(get("/listarFacturas")
                        .contentType(MediaType.APPLICATION_JSON))
                .andExpect(status().isOk())
                .andExpect(jsonPath("$", hasSize(1)))
                .andExpect(jsonPath("$[0].idFactura", is(1)))
                .andExpect(jsonPath("$[0].fechaIngreso", is(fechaPartida[0])))
                .andExpect(jsonPath("$[0].placaFk", is("HHH174")));
    }
}
